Radar memory optimization

+++ radardatasource.h
@@ -3,6 +3,7 @@
 
 #include <stdint.h>
 #include <QObject>
+#include <QtOpenGL>
 
 #if QT_VERSION >= 0x050000
     #include <QtConcurrent/QtConcurrentRun>
@@ -59,17 +60,16 @@ public:
   void finish();
 
 signals:
-  void updateData(uint offset, uint count/*, float* divs*/, float* amps);
+  void updateData(uint offset, uint count, GLubyte* amps);
 
 private:
   bool loadData();
-  bool initWithDummy(/*float* divs, */float* amps);
-  bool loadObserves1(char* filename/*, float* divs*/, float* amps);
-  bool loadObserves2(char* filename/*, float* divs*/, float* amps);
+  bool initWithDummy(GLubyte* amps);
+  bool loadObserves1(char* filename, GLubyte* amps);
+  bool loadObserves2(char* filename, GLubyte* amps);
 
   bool finish_flag;
-  //float file_divs[2][BEARINGS_PER_CYCLE];
-  float file_amps[2][BEARINGS_PER_CYCLE * PELENG_SIZE];
+  GLubyte file_amps[2][BEARINGS_PER_CYCLE * PELENG_SIZE];
   uint  file_curr;
 
   void worker();
